Changed argv index to 1

+++ Aspect.py
@@ -44,12 +44,12 @@ async def on_ready():
     logging.info(f'Logged on as {bot.user.name} id:{bot.user.id} at {datetime.now()}')
     subr: SubredditLinker = SubredditLinker(bot=bot)
     # determine if we need to send reboot confirmation
-    if len(sys.argv) > 2:
-        if len(sys.argv[2]) > len("reboot") and sys.argv[2].startswith("reboot"):
+    if len(sys.argv) > 1:
+        if len(sys.argv[1]) > len("reboot") and sys.argv[2].startswith("reboot"):
             prev_channel = bot.get_channel(id=int(sys.argv[2][len("reboot"):]))
             curr_hash = subprocess.run(["git", "rev-parse", "HEAD"], stdout=subprocess.PIPE, text=True).stdout
             await prev_channel.send(f"Aspect updated to ver: {curr_hash}")
-        elif sys.argv[2] == "reboot":
+        elif sys.argv[1] == "reboot":
             prev_channel = bot.get_channel(id=int(sys.argv[2][len("reboot"):]))
             await prev_channel.send(f"Aspect rebooted.")
